requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20230430
Django==4.2.15
django-admin-numeric-filter==0.1.9
django-celery-beat==2.7.0
django-import-export==4.1.1
django-timezone-field==7.0
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
kombu==5.4.0
multidict==6.0.5
openpyxl==3.2.0b1
pip-system-certs==4.0
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.8
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
